+++ operator/internal/controller/aistore/cluster_controller_test.go
@@ -89,7 +89,6 @@ var _ = Describe("AIStoreController", func() {
 						ProxySpec: aisv1.DaemonSpec{
 							Size: apc.Ptr[int32](1),
 							ServiceSpec: aisv1.ServiceSpec{
-								ServicePort:      intstr.FromInt32(51080),
 								PublicPort:       apc.Ptr(intstr.FromInt32(51081)),
 								IntraControlPort: apc.Ptr(intstr.FromInt32(51082)),
 								IntraDataPort:    apc.Ptr(intstr.FromInt32(51083)),
@@ -99,7 +98,6 @@ var _ = Describe("AIStoreController", func() {
 							DaemonSpec: aisv1.DaemonSpec{
 								Size: apc.Ptr[int32](1),
 								ServiceSpec: aisv1.ServiceSpec{
-									ServicePort:      intstr.FromInt32(51080),
 									PublicPort:       apc.Ptr(intstr.FromInt32(51081)),
 									IntraControlPort: apc.Ptr(intstr.FromInt32(51082)),
 									IntraDataPort:    apc.Ptr(intstr.FromInt32(51083)),